feat: refactor slash menu items to support showing in table cell
