+++ frontend/appflowy_flutter/lib/plugins/document/presentation/editor_plugins/slash_menu/slash_menu_items/math_equation_item.dart
@@ -1,0 +1,42 @@
+import 'package:appflowy/generated/flowy_svgs.g.dart';
+import 'package:appflowy/generated/locale_keys.g.dart';
+import 'package:appflowy/plugins/document/presentation/editor_plugins/base/selectable_svg_widget.dart';
+import 'package:appflowy/plugins/document/presentation/editor_plugins/plugins.dart';
+import 'package:appflowy_editor/appflowy_editor.dart';
+import 'package:easy_localization/easy_localization.dart';
+import 'package:flutter/material.dart';
+
+import 'slash_menu_item_builder.dart';
+
+final _keywords = [
+  'tex',
+  'latex',
+  'katex',
+  'math equation',
+  'formula',
+];
+
+// math equation
+SelectionMenuItem mathEquationSlashMenuItem = SelectionMenuItem.node(
+  getName: () => LocaleKeys.document_slashMenu_name_mathEquation.tr(),
+  keywords: _keywords,
+  nodeBuilder: (editorState, _) => mathEquationNode(),
+  replace: (_, node) => node.delta?.isEmpty ?? false,
+  updateSelection: (editorState, path, __, ___) {
+    WidgetsBinding.instance.addPostFrameCallback((timeStamp) {
+      final mathEquationState =
+          editorState.getNodeAtPath(path)?.key.currentState;
+      if (mathEquationState != null &&
+          mathEquationState is MathEquationBlockComponentWidgetState) {
+        mathEquationState.showEditingDialog();
+      }
+    });
+    return null;
+  },
+  nameBuilder: slashMenuItemNameBuilder,
+  iconBuilder: (editorState, isSelected, style) => SelectableSvgWidget(
+    data: FlowySvgs.slash_menu_icon_math_equation_s,
+    isSelected: isSelected,
+    style: style,
+  ),
+);
